amethyst-server/requirements.txt
Flask
Flask-CORS
tensorflow==2.0.0-alpha0
pillow